Log entries react (#1)
* Align HEREDOC markers

* Remove leading blank lines from specs

* The paper trail config file is now gone as of paper_trail 9.2

* Ignore attempts to set in-place-editable fields to empty values; fixes https://github.com/sul-dlss/exhibits/issues/1183

* Remove unnecessary memoization

* Rewrite a flaky test

* Update spec whitespace to appease rubocop-rspec

* Relax acts-as-taggable-on dependency

* Allow the application catalog config to provide values for custom fields

* Add an easily extendable PageConfigurations object to replace the current hard-coded data hash.

This is intended to allow apps or plugins providing SirTrevor widgets to inject their own configurations into the page.

* :hammer: bump to v2.1.0

* Adding 'no thumb' theming

* Adding 'no thumb' theming

* Merging theming

* Suppressing the toolbar in the details view

* Changed name of viewer to harvard.css to accommodate other CSS changes.

* Mount spotlight at root instead of under /spotlight

* Fix missing newline when appending to config/initializers/assets.rb

Without this fix, the final lines run together as
`Rails.application.config.assets.precompile += %w(
application_nothumb.css )Rails.application.config.assets.precompile +=
%w( application_modern.css )`, and `rake spotlight:server` fails.

Fix missing newline when appending to config/initializers/assets.rb

Put the newline in the right place

* Start prototyping an entrypoint for React

* Get a minimal React entrypoint working

Be sure to run `yarn run build` before checking if the webpack bundle
exists

* Make it easier to update the React assets in development

* Pull in dev config from hul-frontend app

* Move hul-spotlight-frontend src into app/web/

The build config appears to be correct, but there are still a few
missing dependencies

* Update dependencies

There are still some build errors. Just checking dependencies required a
bit of hands-on involvement

* Resolve many (but not all) dependency warnings

There was a lot of noise in the console. This commit gets rid of some of
it.

* Get rid of a lot of package.json files

rm modules/**/package.json did the trick

* Rename $component/$component.js to $component/index.js

Likewise for $container/$container.js

* Finish renaming module files

* Fix autoprefixer and some formatting issues

* Continue culling package.json files and fixing imports

* Start working on other warnings

* Format with prettier

* Add prettier config to package.json script

* Fix some console noise from eslint

* disable build lint; downgrade react-router to 3;

* Finish renaming module files

* Fix warnings in web console, update build

* Add old React Router back

* Update Apollo, get front-end running in Spotlight

* Update config and frontend

* Remove extraneous package.json files

Rename $component/$component.js files to $component/index.js (in keeping
with Node.js convention)

* Move all CSS processing to Webpack

We're now using sass-loader and webpack to handle (S)CSS files.
Also, the build compiles again!

* Use graphql-tag for gql

Updates to files that got pulled in over the weekend

* fix build command;

* disable minification;

* Troubleshoot React loading in production

How do we turn Spotlight on in prod?

* Bypass @babel/polyfill warning

We need to load this for our own use of async functions before
dependencies (which shouldn't be using it anyway) load it